infra/requirements.txt
aws-cdk-lib>=2.207.0
constructs>=10.0.0
cdk-nag==2.36.45
